Switch from float time to double time

+++ src/graphics/graphics_playground.cpp
@@ -182,14 +182,14 @@ int main(int /*argc*/, const char * /*argv*/ [])
     // ImGui
     init_imgui(opengl_version_major, opengl_version_minor, window.window);
 
-    float time = glfwGetTime();
+    double time = glfwGetTime();
 
     // Main loop
     while (!window.should_close())
     {
         // Time
-        float new_time = glfwGetTime();
-        float delta_time = new_time - time;
+        double new_time = glfwGetTime();
+        double delta_time = new_time - time;
         time = new_time;
 
         // Input
