+++ src/graphics/screens/sprite_test_screen.cpp
@@ -36,7 +36,7 @@ SpriteTestScreen::SpriteTestScreen(
 
 SpriteTestScreen::~SpriteTestScreen() {}
 
-void SpriteTestScreen::update(const float delta_time)
+void SpriteTestScreen::update(const double delta_time)
 {
     display_test_dialog("Sprite test", *screens, *screen_names, delta_time, *screen_manager);
     sprite->rotate(1.f * delta_time);
