+++ src/screens/koth_env_screen.h
@@ -30,6 +30,6 @@ class KothEnvScreen : public IScreen
     ~KothEnvScreen();
 
     virtual void draw(Renderer &renderer, bool lightweight = false);
-    void update(const float delta_time);
+    void update(const double delta_time);
 };
 }
